+++ src/server/Modules/CannonClass.lua
@@ -2,28 +2,25 @@ local Cannon = {}
 Cannon.__index = Cannon
 
 -- Services
-local ReplicatedStorage: ReplicatedStorage = game:GetService("ReplicatedStorage")
+local Workspace: Workspace = game:GetService("Workspace")
 
 -- Constructure
 function Cannon.new(config: {})
     local self = setmetatable({}, Cannon)
 
-    self.model = config.model
-
-    self.enterPrompt = self.model:WaitForChild("Barrel"):WaitForChild("EnterPrompt")
-
-    -- Folders
-    local remotes: Folder = ReplicatedStorage:WaitForChild("Remotes")
-
-    -- Remotes
-    local cannonEntered: RemoteEvent = remotes:WaitForChild("CannonEntered")
-
-    -- Events
-    self.enterPrompt.Triggered:Connect(function(player)
-        cannonEntered:FireClient(player, self.model.Name)
-    end)
+    self.model = Workspace:WaitForChild(config.modelName)
+    self.playersInside = {}
 
     return self
+end
+
+-- Class methods
+function Cannon:addPlayer(player: Player)
+    self.playersInside[player] = true
+end
+
+function Cannon:removePlayer(player: Player)
+    self.playersInside[player] = nil
 end
 
 return Cannon