+++ src/server/Modules/CannonConfigs.lua
@@ -1,0 +1,5 @@
+return {
+    MainCannon = {
+        modelName = "MainCannon"
+    }
+}