+++ CHANGELOG.md
@@ -7,12 +7,13 @@
 ### Added
 - `Main.server.lua` - Requires all server modules.
 - `Main.client.lua` - Requires all client modules.
+- `CannonServer.lua` - Contains server logic of cannon.
+- `CannonClient.lua` - Contains client logic of cannon.
 - `CannonClass.lua` - Contains server logic of cannon.
-- `CannonFactory.lua` - Creates cannon instances.
+- `CannonConfigs.lua` - Contains configs of all typs of cannons.
 - `README.md` - Contains game overview and development details.
 - `CHANGELOG.md` - Tracks changes across project versions.
-- `CannonClient.lua` - Contains client logic of cannon.
-  
+
 ---
 
 ## [0.0.1] - 2025-07-27